requirements.txt
asgiref==3.4.1
Django==3.2.6
django-rest-framework-nested==0.0.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
